requirements.txt
Flask==2.2.5
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytest==8.3.5
pytest-cov==5.0.0
playwright==1.53.0
Werkzeug==2.3.8
